fixed bugs, added new functionality and restuctured the code
