+++ package.json
@@ -1,0 +1,25 @@
+{
+  "name": "simulator-webpack",
+  "version": "1.0.0",
+  "description": "",
+  "private": true,
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "webpack",
+    "watch": "watch 'npm run build' ./"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "devDependencies": {
+    "css-loader": "^6.10.0",
+    "npm-watch": "^0.11.0",
+    "style-loader": "^3.3.4",
+    "watch": "^1.0.2",
+    "webpack": "^5.90.3",
+    "webpack-cli": "^5.1.4"
+  },
+  "dependencies": {
+    "@octokit/core": "^6.0.1"
+  }
+}
