+++ src/projectVersion.json
@@ -1,0 +1,3 @@
+{
+    "version": "v1.0.0-alpha"
+}
